requirements.txt
requests
keylogger
pypiwin32
pycryptodome